Minimally feature-complete OLED rendering lua bindings complete.
Default script shows a basic temperature in C and F, CPU usage,
and the current time in local format in various font sizes.

CPU usage is also shown with a 'marching bar graph' on the right
side of the display, and the current 'live' temperature is shown
as a thermometer on the left side.

CLI methods to change the 'image mask' pair applied are pending,
but this is a minimally viable monitoring tool now.