Cleaning up header-generation python scripts
Renaming them to start 'generate_' and converting masks.h to have
a generator script as well so the default_mask.txt is the canonical
and human-readable source code for that header now, much like the
BDF files are the canonical human-readable source code for the
spleen fonts.

Various variable naming cleanups integrated as part of this work.